Make JMSTranslationBundle an optional dependency
The JMSTranslationBundle is only needed if you want to use the bundled extractor. If you don't need the extractor you're being forced to download a bunch of packages that are not really useful and may clash with other dependencies (I am thinking of the php-parser the translation bundle needs). The code doesn't need any more changes because the bundle extension already checks for the presense of the JMSTranslationBundle before registering the extractor.